+++ lib/arachni/rpc/server/framework.rb
@@ -322,8 +322,8 @@ class Framework < ::Arachni::Framework
     # @return   [String]  URL of this instance.
     # @private
     def self_url
-        @self_url ||= @opts.rpc_socket ?
-            @opts.rpc_socket : "#{@opts.rpc_address}:#{@opts.rpc_port}"
+        @self_url ||= @opts.rpc_address ?
+            "#{@opts.rpc_address}:#{@opts.rpc_port}" : @opts.rpc_socket
     end
 
     # @return   [String]  This instance's RPC token.